Meta-d' balance calculation

The expected input is two vectors, one for responses to each stimulus, encoding the observers response and confidence. For example, for two stimului labelled A and B, with three confidence ratings, participants could respond to stimulus A as follows: Response: A, rating: 3, count: 60 Response: A, rating: 2, count: 30 Response: A, rating: 1, count: 10 Response: B, rating: 1, count: 7 Response: B, rating: 2, count: 4 Response: B, rating: 3, count: 1

fit_meta_d_bal(nr_s1, nr_s2, s = 1, add_constant = FALSE)

Arguments

nr_s1

Responses to S1 stimulus. See below for advice.

nr_s2

Responses to S2 stimulus. See below for advice.

s

Ratio of standard deviations for the S1 and S2 stimulus. Defaults to 1.

add_constant

Add a small constant to all cells to adjust for boundary issues, and for consistency with the use of this method with other meta-d measures. Note: default for this is FALSE.

Details

The appropriate vector would be nr_s1 <- c(60,30,10,7,4,1)

For stimulus B, we would have the respective vector for responses to stimulus B, eg: Response: A, rating: 3, count: 4 Response: A, rating: 2, count: 6 Response: A, rating: 1, count: 11 Response: B, rating: 1, count: 13 Response: B, rating: 2, count: 23 Response: B, rating: 3, count: 61

nr_s2 <- c(4,6,11,13,23,61)

References

Barrett, Dienes, & Seth (2013). Measures of metacognition on signal-detection theoretic models. Psychol Methods, 18. http://dx.doi.org/10.1037/a0033268

See also

Author

Adam Barrett. Ported to R by Matt Craddock matt@mattcraddock.com

Examples

nr_s1 <- c(60,30,10,7,4,1)
nr_s2 <- c(4,6,11,13,23,61)
fit_meta_d_bal(nr_s1, nr_s2)
#>    d_prime meta_d_bal meta_d_ratio  meta_d_diff stable       htp       ftp
#> 1 2.165011   2.165243     1.000107 0.0002318857      1 0.8358748 0.2336946
#>         htm      ftm       t1c1
#> 1 0.9390267 0.125132 -0.1593614